Write static file instead
