+++ packages/cli/src/commands/start.ts
@@ -6,7 +6,7 @@ import { Command } from '@n8n/decorators';
 import { Container } from '@n8n/di';
 import glob from 'fast-glob';
 import { createReadStream, createWriteStream, existsSync } from 'fs';
-import { mkdir } from 'fs/promises';
+import { writeFile, mkdir } from 'fs/promises';
 import { jsonParse, randomString, type IWorkflowExecutionDataProcess } from 'n8n-workflow';
 import path from 'path';
 import replaceStream from 'replacestream';
@@ -16,7 +16,7 @@ import { z } from 'zod';
 import { ActiveExecutions } from '@/active-executions';
 import { ActiveWorkflowManager } from '@/active-workflow-manager';
 import config from '@/config';
-import { EDITOR_UI_DIST_DIR } from '@/constants';
+import { EDITOR_UI_DIST_DIR, N8N_VERSION } from '@/constants';
 import { FeatureNotLicensedError } from '@/errors/feature-not-licensed.error';
 import { MessageEventBus } from '@/eventbus/message-event-bus/message-event-bus';
 import { EventService } from '@/events/event.service';
@@ -161,9 +161,28 @@ export class Start extends BaseCommand<z.infer<typeof flagsSchema>> {
 			}
 		};
 
-		await compileFile('index.html');
+		const generateConfigFile = async () => {
+			const configPath = path.join(staticCacheDir, 'config.js');
+
+			const feSentryConfig = {
+				dsn: this.globalConfig.sentry.frontendDsn,
+				// eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+				environment: process.env.ENVIRONMENT || 'development',
+				serverName: process.env.DEPLOYMENT_NAME,
+				release: `n8n@${N8N_VERSION}`,
+			};
+
+			const configContent = [
+				`window.BASE_PATH = '${n8nPath}';`,
+				`window.REST_ENDPOINT = '${this.globalConfig.endpoints.rest}';`,
+				`window.sentry = ${JSON.stringify(feSentryConfig)};`,
+			].join('\n');
+
+			await writeFile(configPath, configContent, 'utf-8');
+		};
+
 		const files = await glob('**/*.{css,js}', { cwd: EDITOR_UI_DIST_DIR });
-		await Promise.all(files.map(compileFile));
+		await Promise.all([generateConfigFile(), compileFile('index.html'), ...files.map(compileFile)]);
 	}
 
 	async init() {
